Add _strict argument to Parameter call
This is so the client (SpineModel) doesn't need to check first
if the parameter is specified. They can just pass _stric=false
instead and the function will return nothing instead of raising.

We could also consider use a _default argument to specify a value
to return instead of nothing, but maybe later.