Let TimeSlice roll
The start and end_ of a TimeSlice can now be modified using roll!
This is intended to be used in a rolling horizon scheme where
the same problem is solved many times after rolling forward in
time a fixed amount.

We don't want dictionary entries to lose track of TimeSlices so
we 'hack' the Base.objectid method for TimeSlices.

To retrieve the start of a TimeSlice one needs to use `t.start[]`
instead of just `t.start`. But better is to use the provided method
`start(t)`.