Introduce `Call`
`Call` is a simple data structure to represent a function call.
A `Call` can be `realize`d which means actually performing the call
to retrieve the result.

The idea is to use `Call` in `SpineModel.jl` to keep track of
convenience function calls in constraint code, thus enabling
automatic model updating in rolling horizon optimizations.