Try to optimize `Call` api
The Call api is unbearably slow, this commit is a WIP towards making
it faster. We manage to reduce allocations but it feels like we need
to keep working.

For the record, the current strategy is to make `Call` an abstract
type, and then provide optimized (?) subtypes for Parameter,
general Function, and scalar.