Introduce `add_objects!` and `add_relationships!`
These methods replace `push_default_object!` and
`push_default_relationship!` respectively.
The intention is to try and make it efficient to add a large number
of new objects or relationships.

Also we don't need to copy default values as the convenience
function is able to find them from the `default_values` field.

We also move the check for existence inside the functions so the
caller doesn't need to bother. `setdiff!` comes in handy.