Refactor Map and MapParameterValue
Re #32

The idea is to now support calling parameters with the `s` keyword
argument that corresponds to a stochastic scenario `Object`.

This is complemented by a new way of parsing "map" parameter values
from the db. Basically, a `spinedb_api.Map` is always converted
into a `SpineInterface.Map`, *except* when the "index_type" is
"date_time", in which case it is converted to a
`SpineInterface.TimeSeries`.

Also, we use a dictionary to store `Map`s indices and values in a
more "performant" way.

When a `MapParameterValue` is called with the `s` argument, the
function looks for a key in the dictionary corresponding to `s.name`.
It everything goes well, it will find a `TimeSeriesParameterValue`
that it will proceed to call with the `t::TimeSlice` argument.

A test set has been added to make sure this basic setup works.