#32 Make the `prefix` tolerant of additional arguments.
Changes the behaviour of the `MapParameterValue` so that instead of returning `nothing` if a `prefix` key is not found, it simply continues to search for the other remaining keys. This has the upside that parameter calls in SpineOpt should become tolerant of full *updating stochastic time series*, *non-updating stochastic time series*, as well as *updating non-stochastic time series* all with the same parameter call. Hopefully this doesn't cause too much trouble when a `Map` is returned instead of `nothing` on wrong calls...