Get rid of `ObjectGroup` and use `Object` also for groups
It's better to manage a single type, because of how things work
down the line, especially when creating NamedTuples with Object values.
In particular, an array of NamedTuples will only have the 'right'
type if all the values for a given key have the same type.
So, if for some of the elements, the key `:node` has a value of type
`Object` and for some others `ObjectGroup`, julia won't be able to
determine the 'right' element type for the array and some of our
overrides in SpineOpt won't work.

We also introduce `members` to replace `expand`, and `groups` to
return an array of all `Object`s where an `Object` passed as
argument happens to be a member.