`MapParameterValue` now returns the latest `DateTime` key if called with a `DateTime` key.
Previously, the keys had to match in order for the `MapParameterValue` to find any values. Now, if the `MapParameterValue` indices are `DateTime`s and it's called with a `DateTime` value, the `MapParameterValue` returns the value from the latest valid `DateTime` key. This is due to a requirement from updating time series data. If we want to generalize this more, we could implement similar behaviour for other key types that form a "number line".