Form *prefix* from unhandled kwargs in the call to `MapParameterValue`
The idea here is that the user of `MapParameterValue` can use a liberal
notation, so any keyword argument that's not `t` and `i` (i.e.,
the ones reserved specifically by other callables) are taken as the
map *prefix*. The `prefix` keyword argument is no longer handled.

So, if the user calls `map(whatever=s, whocares=h, t=t, i=i)` then
the *prefix* will automatically be `(s, h)`.