Remove `Pair` wrapper in `MapParameterValue` calls
Also, add a couple of stress tests where we try to confuse the
`MapParameterValue` callable by sneaking dummy kwargs.
The point is the callable is quite liberal at ignoring
keyword arguments that are wrong so to say, or at least not findable
in the Map. Anyways, seems good for now.