Replace ParameterCall by ParameterValueCall
The thing is, when we created the ParameterCall, we looked up the
parameter value inside the Parameter to determine whether or not
it was time-varying. And then, when calling `realize` on the
`Call`, we looked up the parameter value again, sort of ignoring
we already did.

I think this had to do with how we managed the keyword arguments
in the Parameter callable. Since this recently changed, now we're
able to look up the parameter value only once at the moment of
creating the Call, and just call the parameter value when
`realize`ing it.