When rolling time slice maps, also roll the time slices in the map
This is just a detail, but it doesn't seem correct to just roll
the start and end_ of a time slice map, without rolling the time slices
too. That would result in the map being corrupted so to say.

So now we roll the time slices together with the start and end_.
This change impacts SpineOpt.roll_temporal_structure so I'm also
bumping the version patch here.