Fix write_parameters so it always goes
The idea here is that write_parameters shouldn't fail because the
given url is not 'ready'. We now (i) create one if not existent
(which we already did, but now we use the `create` keyword
argument to `DiffDatabaseMapping`, (ii) upgrade an existing one
to the latest revision, and (iii) inject the Base alternative just
in case.