Extend the Map API
We were missing a key element which is a way to write Map back
to the db using write_parameters.

This required a few steps:

1. Stop supporting duplicate indices in Maps from the db. spinedb_api
doesn't check that indices are unique, but we kind of need it here
otherwise we don't know how to uniquely retrieve a value. Previously
we had a system where each index was mapped to an Array of values
but that didn't make much sense. Now we raise a warning and keep only
one of the duplicated indices (I think it's the first one).
2. Fix the Map tests
3. Refactor maximum_parameter_value slightly and provide a test.