Tests for parameter value range, as well as object class and relationship class counts are only performed when applicable.
This makes the testing output more compact for failing tests. E.g. if a parameter value has a type mismatch, the range test will likely fail or even throw an error, so that step if skipped. Similarly, counting object and relationship class overlaps makes no sense if there is no overlap to begin with.