Support parameter values with Symbols
As if they were strings. This is because SpineInterface.Map() converts
string indices to symbols, so calling the convenience function with
the output of `stochastic_scenario` works. Maybe that's not such a
great idea and we should handle the Symbol in the parameter value
access function. I will try that next, this is so things work quickly
now.