Move timedata_operation to api.jl
util.jl is intended for internal utility, ie helper functions used
by other files in the package. We don't export anything from here
and we name things with a leading underscore to stress this fact.

On the contrary, for stuff that we export we use api.jl
(even if one could think of it as 'utility')