Parse null JSON in time series and arrays as NaN
NaN values in time series and arrays are imported as JSON null when
using the db server, because we need to serialize the data using JSON
and JSON doesn't know NaN (even though python's json does).

The result is we cannot write NaN into the db unless we tweak our JSON
encoding here and I don't have time for that right now. This makes
sense anyways, and gets around the issue too.