Replace RelationshipClass.lookup_cache by row_map
We were memoizing calls to a RelationshipClass using a Dict that
we called lookup_cache.
This approach has a few drawbacks:
- The benefit only comes the second time you call the function.
- You may be storing too much data because the arguments are pretty
  arbitrary - large memory footprint.
- The cache is not built optimally. Each time we call the function,
  we perform a full pass over the entities to build the cache
  for a single tuple of arguments.

Now we precompute an equivalent of the old lookup cache via a single
pass when creating the RelationshipClass object.
It does more work upfront but it helps with all three issues listed
above.